dialog 3
Create Dialog - Dialog must consist of an EditText for accepting the
user input (number) with
„OK‟ & „CANCEL‟ Buttons. On the click of OK Button collect the number
entered in the EditText
and pass the same to AsyncTask.
Create AsyncTask - AsyncTask must start & execute for x secs(x is the
number passed from the
click on OK Button) which is passed as the argument to the AsyncTask,
with the ProgressDialog
being display to the user during the AsyncTask execution. Upon
successful completion of
execution, from AsyncTask‟s onPostExecute() method create an AlertDialog
with the title
“Successful” and text “Completed” along with „OK‟ Button. Upon the click
on the OK Button
the AlertDialog must be dismissed.
On the click of CANCEL Button the Dialog must be closed